pyproject.toml
[project]
name = "tses-assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autopep8>=2.3.2",
    "boto3>=1.40.55",
    "botocore>=1.40.55",
    "celery>=5.5.3",
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "colorlog>=6.10.1",
    "django>=5.2.7",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.0.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-import-export>=4.3.12",
    "django-ipware>=7.0.1",
    "django-jazzmin>=3.0.1",
    "django-redis>=6.0.0",
    "django-storages>=1.14.6",
    "django-widget-tweaks>=1.5.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "docutils>=0.22.2",
    "drf-spectacular>=0.28.0",
    "drf-standardized-errors>=0.15.0",
    "email-validator>=2.3.0",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "locust>=2.42.0",
    "markdown>=3.9",
    "markupsafe>=3.0.3",
    "pillow>=12.0.0",
    "psutil>=7.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pyotp>=2.9.0",
    "python-dateutil>=2.9.0.post0",
    "redis>=6.4.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "pylint>=3.3.6",
    "pylint-django>=2.6.1",
    "ruff>=0.11.2",
]
test = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.0",
    "pytest-django>=4.10.0",
]

[tool.isort]
profile = "black"
combine_as_imports = true
include_trailing_comma = true
line_length = 120

[tool.ruff]
line-length = 120
